% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenCausalEst.R
\name{Gen_Estimand}
\alias{Gen_Estimand}
\title{Causal inference via general estimands: the weighted ATE, weighted ATT and weighted ATC}
\usage{
Gen_Estimand(
  A,
  Y,
  X,
  beta = FALSE,
  v = NA,
  trim = FALSE,
  trim.alpha = 0.05,
  trun = FALSE,
  trun.alpha = 0.05,
  boot = TRUE,
  n.boot = 500,
  conf.level = 0.05
)
}
\arguments{
\item{A}{vector of the treatment variable (binary, valued from 0 and 1)}

\item{Y}{vector of the outcome variable (can be various types: continous, categorical, etc., but must be numerical)}

\item{X}{matrix of the covariates that included into the propensity score model}

\item{beta}{whether to include estimands from beta family weights; if so, the v parameter below is required to be specified}

\item{v}{model parameter for beta weights}

\item{trim}{whether to include trimming; if so, the trim.alpha parameter below needs to be specified}

\item{trim.alpha}{the trimming threshold; default is .05}

\item{trun}{whether to include truncation; if so, the trun.alpha parameter below needs to be specified}

\item{trun.alpha}{the truncation threshold; default is .05}

\item{boot}{whether to implement bootstrap for variance estimation; default is TRUE}

\item{n.boot}{number of bootstrap, if boot==TRUE; default is 500}

\item{conf.level}{level of confidence interval; default is .05 (for 0.95 confidence interval, using normal approximation)}
}
\value{
the function returns estimated causal estimands using observational data;
        estimands defined by constant weights, overlap weights, matching weights and entropy weights will always be returned;
        whether to return estimands defined by trimming, truncation, and beta weights can be specified by user
}
\description{
Causal inference via general estimands: the weighted ATE, weighted ATT and weighted ATC
}
